statistics: add analysis job factory

refacor: move to priorityqueue_test


refactor: better time window


refactor: better parition name and id


docs: add comments back


docs: add comments back


docs: add comments back


tests: add more cases


refactor: make lint happy


docs: mark it is not thread safe


refactor: rename


fix
