+++ pkg/statistics/handle/autoanalyze/refresher/refresher.go
@@ -23,7 +23,6 @@ import (
 	"github.com/pingcap/tidb/pkg/sessionctx"
 	"github.com/pingcap/tidb/pkg/sessionctx/sysproctrack"
 	"github.com/pingcap/tidb/pkg/sessionctx/variable"
-	"github.com/pingcap/tidb/pkg/statistics"
 	"github.com/pingcap/tidb/pkg/statistics/handle/autoanalyze/exec"
 	"github.com/pingcap/tidb/pkg/statistics/handle/autoanalyze/priorityqueue"
 	"github.com/pingcap/tidb/pkg/statistics/handle/lockstats"
@@ -32,16 +31,7 @@ import (
 	statsutil "github.com/pingcap/tidb/pkg/statistics/handle/util"
 	"github.com/pingcap/tidb/pkg/util"
 	"github.com/pingcap/tidb/pkg/util/intest"
-	"github.com/pingcap/tidb/pkg/util/timeutil"
-	"github.com/tikv/client-go/v2/oracle"
 	"go.uber.org/zap"
-)
-
-const (
-	// unanalyzedTableDefaultChangePercentage is the default change percentage of unanalyzed table.
-	unanalyzedTableDefaultChangePercentage = 1
-	// unanalyzedTableDefaultLastUpdateDuration is the default last update duration of unanalyzed table.
-	unanalyzedTableDefaultLastUpdateDuration = -30 * time.Minute
 )
 
 // Refresher provides methods to refresh stats info.
@@ -50,7 +40,7 @@ type Refresher struct {
 	statsHandle    statstypes.StatsHandle
 	sysProcTracker sysproctrack.Tracker
 	// This will be refreshed every time we rebuild the priority queue.
-	autoAnalysisTimeWindow
+	autoAnalysisTimeWindow priorityqueue.AutoAnalysisTimeWindow
 
 	// Jobs is the priority queue of analysis jobs.
 	// Exported for testing purposes.
@@ -84,7 +74,7 @@ func (r *Refresher) UpdateConcurrency() {
 
 // AnalyzeHighestPriorityTables picks tables with the highest priority and analyzes them.
 func (r *Refresher) AnalyzeHighestPriorityTables() bool {
-	if !r.autoAnalysisTimeWindow.isWithinTimeWindow(time.Now()) {
+	if !r.autoAnalysisTimeWindow.IsWithinTimeWindow(time.Now()) {
 		return false
 	}
 
@@ -182,11 +172,8 @@ func (r *Refresher) RebuildTableAnalysisJobQueue() error {
 			}
 			// We will check it again when we try to execute the job.
 			// So store the time window for later use.
-			r.autoAnalysisTimeWindow = autoAnalysisTimeWindow{
-				start: start,
-				end:   end,
-			}
-			if !r.autoAnalysisTimeWindow.isWithinTimeWindow(time.Now()) {
+			r.autoAnalysisTimeWindow = priorityqueue.NewAutoAnalysisTimeWindow(start, end)
+			if !r.autoAnalysisTimeWindow.IsWithinTimeWindow(time.Now()) {
 				return nil
 			}
 			calculator := priorityqueue.NewPriorityCalculator()
@@ -204,11 +191,13 @@ func (r *Refresher) RebuildTableAnalysisJobQueue() error {
 				return err
 			}
 
+			jobFactory := priorityqueue.NewAnalysisJobFactory(sctx, autoAnalyzeRatio, currentTs)
+
 			dbs := is.AllSchemaNames()
 			for _, db := range dbs {
 				// Sometimes the tables are too many. Auto-analyze will take too much time on it.
 				// so we need to check the available time.
-				if !r.autoAnalysisTimeWindow.isWithinTimeWindow(time.Now()) {
+				if !r.autoAnalysisTimeWindow.IsWithinTimeWindow(time.Now()) {
 					return nil
 				}
 				// Ignore the memory and system database.
@@ -231,37 +220,13 @@ func (r *Refresher) RebuildTableAnalysisJobQueue() error {
 						continue
 					}
 					pi := tblInfo.GetPartitionInfo()
-					pushJobFunc := func(job priorityqueue.AnalysisJob) {
-						if job == nil {
-							return
-						}
-						// Calculate the weight of the job.
-						weight := calculator.CalculateWeight(job)
-						// We apply a penalty to larger tables, which can potentially result in a negative weight.
-						// To prevent this, we filter out any negative weights. Under normal circumstances, table sizes should not be negative.
-						if weight <= 0 {
-							statslogutil.SingletonStatsSamplerLogger().Warn(
-								"Table gets a negative weight",
-								zap.Float64("weight", weight),
-								zap.Stringer("job", job),
-							)
-						}
-						job.SetWeight(weight)
-						// Push the job onto the queue.
-						r.Jobs.Push(job)
-					}
-					// No partitions, analyze the whole table.
 					if pi == nil {
-						job := CreateTableAnalysisJob(
-							sctx,
+						job := jobFactory.CreateNonPartitionedTableAnalysisJob(
 							db.O,
 							tblInfo,
 							r.statsHandle.GetTableStatsForAutoAnalyze(tblInfo),
-							autoAnalyzeRatio,
-							currentTs,
 						)
-						pushJobFunc(job)
-						// Skip the rest of the loop.
+						r.pushJob(job, calculator)
 						continue
 					}
 
@@ -272,33 +237,27 @@ func (r *Refresher) RebuildTableAnalysisJobQueue() error {
 							partitionDefs = append(partitionDefs, def)
 						}
 					}
-					partitionStats := getPartitionStats(r.statsHandle, tblInfo, partitionDefs)
+					partitionStats := priorityqueue.GetPartitionStats(r.statsHandle, tblInfo, partitionDefs)
 					// If the prune mode is static, we need to analyze every partition as a separate table.
 					if pruneMode == variable.Static {
 						for pIDAndName, stats := range partitionStats {
-							job := CreateStaticPartitionAnalysisJob(
-								sctx,
+							job := jobFactory.CreateStaticPartitionAnalysisJob(
 								db.O,
 								tblInfo,
 								pIDAndName.ID,
 								pIDAndName.Name,
 								stats,
-								autoAnalyzeRatio,
-								currentTs,
 							)
-							pushJobFunc(job)
+							r.pushJob(job, calculator)
 						}
 					} else {
-						job := createTableAnalysisJobForPartitions(
-							sctx,
+						job := jobFactory.CreateDynamicPartitionedTableAnalysisJob(
 							db.O,
 							tblInfo,
 							r.statsHandle.GetPartitionStatsForAutoAnalyze(tblInfo, tblInfo.ID),
 							partitionStats,
-							autoAnalyzeRatio,
-							currentTs,
 						)
-						pushJobFunc(job)
+						r.pushJob(job, calculator)
 					}
 				}
 			}
@@ -311,6 +270,25 @@ func (r *Refresher) RebuildTableAnalysisJobQueue() error {
 	}
 
 	return nil
+}
+
+func (r *Refresher) pushJob(job priorityqueue.AnalysisJob, calculator *priorityqueue.PriorityCalculator) {
+	if job == nil {
+		return
+	}
+	// We apply a penalty to larger tables, which can potentially result in a negative weight.
+	// To prevent this, we filter out any negative weights. Under normal circumstances, table sizes should not be negative.
+	weight := calculator.CalculateWeight(job)
+	if weight <= 0 {
+		statslogutil.SingletonStatsSamplerLogger().Warn(
+			"Table gets a negative weight",
+			zap.Float64("weight", weight),
+			zap.Stringer("job", job),
+		)
+	}
+	job.SetWeight(weight)
+	// Push the job onto the queue.
+	r.Jobs.Push(job)
 }
 
 // WaitAutoAnalyzeFinishedForTest waits for the auto analyze job to be finished.
@@ -330,363 +308,10 @@ func (r *Refresher) Close() {
 	r.worker.Stop()
 }
 
-// CreateTableAnalysisJob creates a TableAnalysisJob for the physical table.
-func CreateTableAnalysisJob(
-	sctx sessionctx.Context,
-	tableSchema string,
-	tblInfo *model.TableInfo,
-	tblStats *statistics.Table,
-	autoAnalyzeRatio float64,
-	currentTs uint64,
-) priorityqueue.AnalysisJob {
-	if !tblStats.IsEligibleForAnalysis() {
-		return nil
-	}
-
-	tableStatsVer := sctx.GetSessionVars().AnalyzeVersion
-	statistics.CheckAnalyzeVerOnTable(tblStats, &tableStatsVer)
-
-	changePercentage := CalculateChangePercentage(tblStats, autoAnalyzeRatio)
-	tableSize := calculateTableSize(tblStats)
-	lastAnalysisDuration := GetTableLastAnalyzeDuration(tblStats, currentTs)
-	indexes := CheckIndexesNeedAnalyze(tblInfo, tblStats)
-
-	// No need to analyze.
-	// We perform a separate check because users may set the auto analyze ratio to 0,
-	// yet still wish to analyze newly added indexes and tables that have not been analyzed.
-	if changePercentage == 0 && len(indexes) == 0 {
-		return nil
-	}
-
-	job := priorityqueue.NewNonPartitionedTableAnalysisJob(
-		tableSchema,
-		tblInfo.Name.O,
-		tblInfo.ID,
-		indexes,
-		tableStatsVer,
-		changePercentage,
-		tableSize,
-		lastAnalysisDuration,
-	)
-
-	return job
-}
-
-// CreateStaticPartitionAnalysisJob creates a TableAnalysisJob for the static partition.
-func CreateStaticPartitionAnalysisJob(
-	sctx sessionctx.Context,
-	tableSchema string,
-	globalTblInfo *model.TableInfo,
-	partitionID int64,
-	partitionName string,
-	partitionStats *statistics.Table,
-	autoAnalyzeRatio float64,
-	currentTs uint64,
-) priorityqueue.AnalysisJob {
-	if !partitionStats.IsEligibleForAnalysis() {
-		return nil
-	}
-
-	tableStatsVer := sctx.GetSessionVars().AnalyzeVersion
-	statistics.CheckAnalyzeVerOnTable(partitionStats, &tableStatsVer)
-
-	changePercentage := CalculateChangePercentage(partitionStats, autoAnalyzeRatio)
-	tableSize := calculateTableSize(partitionStats)
-	lastAnalysisDuration := GetTableLastAnalyzeDuration(partitionStats, currentTs)
-	indexes := CheckIndexesNeedAnalyze(globalTblInfo, partitionStats)
-
-	// No need to analyze.
-	// We perform a separate check because users may set the auto analyze ratio to 0,
-	// yet still wish to analyze newly added indexes and tables that have not been analyzed.
-	if changePercentage == 0 && len(indexes) == 0 {
-		return nil
-	}
-
-	job := priorityqueue.NewStaticPartitionTableAnalysisJob(
-		tableSchema,
-		globalTblInfo.Name.O,
-		globalTblInfo.ID,
-		partitionName,
-		partitionID,
-		indexes,
-		tableStatsVer,
-		changePercentage,
-		tableSize,
-		lastAnalysisDuration,
-	)
-
-	return job
-}
-
-// CalculateChangePercentage calculates the change percentage of the table
-// based on the change count and the analysis count.
-func CalculateChangePercentage(
-	tblStats *statistics.Table,
-	autoAnalyzeRatio float64,
-) float64 {
-	if !tblStats.IsAnalyzed() {
-		return unanalyzedTableDefaultChangePercentage
-	}
-
-	// Auto analyze based on the change percentage is disabled.
-	// However, this check should not affect the analysis of indexes,
-	// as index analysis is still needed for query performance.
-	if autoAnalyzeRatio == 0 {
-		return 0
-	}
-
-	tblCnt := float64(tblStats.RealtimeCount)
-	if histCnt := tblStats.GetAnalyzeRowCount(); histCnt > 0 {
-		tblCnt = histCnt
-	}
-	res := float64(tblStats.ModifyCount) / tblCnt
-	if res > autoAnalyzeRatio {
-		return res
-	}
-
-	return 0
-}
-
-func calculateTableSize(
-	tblStats *statistics.Table,
-) float64 {
-	tblCnt := float64(tblStats.RealtimeCount)
-	colCnt := float64(tblStats.ColAndIdxExistenceMap.ColNum())
-	intest.Assert(colCnt != 0, "Column count should not be 0")
-
-	return tblCnt * colCnt
-}
-
-// GetTableLastAnalyzeDuration gets the duration since the last analysis of the table.
-func GetTableLastAnalyzeDuration(
-	tblStats *statistics.Table,
-	currentTs uint64,
-) time.Duration {
-	lastTime := findLastAnalyzeTime(tblStats, currentTs)
-	currentTime := oracle.GetTimeFromTS(currentTs)
-
-	// Calculate the duration since last analyze.
-	return currentTime.Sub(lastTime)
-}
-
-// findLastAnalyzeTime finds the last analyze time of the table.
-// It uses `LastUpdateVersion` to find the last analyze time.
-// The `LastUpdateVersion` is the version of the transaction that updates the statistics.
-// It always not null(default 0), so we can use it to find the last analyze time.
-func findLastAnalyzeTime(
-	tblStats *statistics.Table,
-	currentTs uint64,
-) time.Time {
-	// Table is not analyzed, compose a fake version.
-	if !tblStats.IsAnalyzed() {
-		phy := oracle.GetTimeFromTS(currentTs)
-		return phy.Add(unanalyzedTableDefaultLastUpdateDuration)
-	}
-	return oracle.GetTimeFromTS(tblStats.LastAnalyzeVersion)
-}
-
-// CheckIndexesNeedAnalyze checks if the indexes of the table need to be analyzed.
-func CheckIndexesNeedAnalyze(
-	tblInfo *model.TableInfo,
-	tblStats *statistics.Table,
-) []string {
-	// If table is not analyzed, we need to analyze whole table.
-	// So we don't need to check indexes.
-	if !tblStats.IsAnalyzed() {
-		return nil
-	}
-
-	indexes := make([]string, 0, len(tblInfo.Indices))
-	// Check if missing index stats.
-	for _, idx := range tblInfo.Indices {
-		if idxStats := tblStats.GetIdx(idx.ID); idxStats == nil && !tblStats.ColAndIdxExistenceMap.HasAnalyzed(idx.ID, true) && idx.State == model.StatePublic {
-			indexes = append(indexes, idx.Name.O)
-		}
-	}
-
-	return indexes
-}
-
-func createTableAnalysisJobForPartitions(
-	sctx sessionctx.Context,
-	tableSchema string,
-	tblInfo *model.TableInfo,
-	tblStats *statistics.Table,
-	partitionStats map[PartitionIDAndName]*statistics.Table,
-	autoAnalyzeRatio float64,
-	currentTs uint64,
-) priorityqueue.AnalysisJob {
-	if !tblStats.IsEligibleForAnalysis() {
-		return nil
-	}
-
-	// TODO: figure out how to check the table stats version correctly for partitioned tables.
-	tableStatsVer := sctx.GetSessionVars().AnalyzeVersion
-	statistics.CheckAnalyzeVerOnTable(tblStats, &tableStatsVer)
-
-	averageChangePercentage, avgSize, minLastAnalyzeDuration, partitionNames := CalculateIndicatorsForPartitions(
-		tblStats,
-		partitionStats,
-		autoAnalyzeRatio,
-		currentTs,
-	)
-	partitionIndexes := CheckNewlyAddedIndexesNeedAnalyzeForPartitionedTable(
-		tblInfo,
-		partitionStats,
-	)
-	// No need to analyze.
-	// We perform a separate check because users may set the auto analyze ratio to 0,
-	// yet still wish to analyze newly added indexes and tables that have not been analyzed.
-	if len(partitionNames) == 0 && len(partitionIndexes) == 0 {
-		return nil
-	}
-
-	job := priorityqueue.NewDynamicPartitionedTableAnalysisJob(
-		tableSchema,
-		tblInfo.Name.O,
-		tblInfo.ID,
-		partitionNames,
-		partitionIndexes,
-		tableStatsVer,
-		averageChangePercentage,
-		avgSize,
-		minLastAnalyzeDuration,
-	)
-
-	return job
-}
-
-// CalculateIndicatorsForPartitions calculates the average change percentage,
-// average size and average last analyze duration for the partitions that meet the threshold.
-// Change percentage is the ratio of the number of modified rows to the total number of rows.
-// Size is the product of the number of rows and the number of columns.
-// Last analyze duration is the duration since the last analyze.
-func CalculateIndicatorsForPartitions(
-	globalStats *statistics.Table,
-	partitionStats map[PartitionIDAndName]*statistics.Table,
-	autoAnalyzeRatio float64,
-	currentTs uint64,
-) (
-	avgChange float64,
-	avgSize float64,
-	avgLastAnalyzeDuration time.Duration,
-	partitionNames []string,
-) {
-	totalChangePercent := 0.0
-	totalSize := 0.0
-	count := 0.0
-	partitionNames = make([]string, 0, len(partitionStats))
-	cols := float64(globalStats.ColAndIdxExistenceMap.ColNum())
-	intest.Assert(cols != 0, "Column count should not be 0")
-	totalLastAnalyzeDuration := time.Duration(0)
-
-	for pIDAndName, tblStats := range partitionStats {
-		changePercent := CalculateChangePercentage(tblStats, autoAnalyzeRatio)
-		// Skip partition analysis if it doesn't meet the threshold, stats are not yet loaded,
-		// or the auto analyze ratio is set to 0 by the user.
-		if changePercent == 0 {
-			continue
-		}
-
-		totalChangePercent += changePercent
-		// size = count * cols
-		totalSize += float64(tblStats.RealtimeCount) * cols
-		lastAnalyzeDuration := GetTableLastAnalyzeDuration(tblStats, currentTs)
-		totalLastAnalyzeDuration += lastAnalyzeDuration
-		partitionNames = append(partitionNames, pIDAndName.Name)
-		count++
-	}
-	if len(partitionNames) == 0 {
-		return 0, 0, 0, partitionNames
-	}
-
-	avgChange = totalChangePercent / count
-	avgSize = totalSize / count
-	avgLastAnalyzeDuration = totalLastAnalyzeDuration / time.Duration(count)
-
-	return avgChange, avgSize, avgLastAnalyzeDuration, partitionNames
-}
-
-// CheckNewlyAddedIndexesNeedAnalyzeForPartitionedTable checks if the indexes of the partitioned table need to be analyzed.
-// It returns a map from index name to the names of the partitions that need to be analyzed.
-// NOTE: This is only for newly added indexes.
-func CheckNewlyAddedIndexesNeedAnalyzeForPartitionedTable(
-	tblInfo *model.TableInfo,
-	partitionStats map[PartitionIDAndName]*statistics.Table,
-) map[string][]string {
-	partitionIndexes := make(map[string][]string, len(tblInfo.Indices))
-
-	for _, idx := range tblInfo.Indices {
-		// No need to analyze the index if it's not public.
-		if idx.State != model.StatePublic {
-			continue
-		}
-
-		// Find all the partitions that need to analyze this index.
-		names := make([]string, 0, len(partitionStats))
-		for pIDAndName, tblStats := range partitionStats {
-			if idxStats := tblStats.GetIdx(idx.ID); idxStats == nil && !tblStats.ColAndIdxExistenceMap.HasAnalyzed(idx.ID, true) {
-				names = append(names, pIDAndName.Name)
-			}
-		}
-
-		if len(names) > 0 {
-			partitionIndexes[idx.Name.O] = names
-		}
-	}
-
-	return partitionIndexes
-}
-
 func getStartTs(sctx sessionctx.Context) (uint64, error) {
 	txn, err := sctx.Txn(true)
 	if err != nil {
 		return 0, err
 	}
 	return txn.StartTS(), nil
-}
-
-// PartitionIDAndName is a struct that contains the ID and name of a partition.
-// Exported for testing purposes. Do not use it in other packages.
-type PartitionIDAndName struct {
-	Name string
-	ID   int64
-}
-
-func getPartitionStats(
-	statsHandle statstypes.StatsHandle,
-	tblInfo *model.TableInfo,
-	defs []model.PartitionDefinition,
-) map[PartitionIDAndName]*statistics.Table {
-	partitionStats := make(map[PartitionIDAndName]*statistics.Table, len(defs))
-
-	for _, def := range defs {
-		stats := statsHandle.GetPartitionStatsForAutoAnalyze(tblInfo, def.ID)
-		// Ignore the partition if it's not ready to analyze.
-		if !stats.IsEligibleForAnalysis() {
-			continue
-		}
-		d := PartitionIDAndName{
-			ID:   def.ID,
-			Name: def.Name.O,
-		}
-		partitionStats[d] = stats
-	}
-
-	return partitionStats
-}
-
-// autoAnalysisTimeWindow is a struct that contains the start and end time of the auto analyze time window.
-type autoAnalysisTimeWindow struct {
-	start time.Time
-	end   time.Time
-}
-
-// isWithinTimeWindow checks if the current time is within the time window.
-// If the auto analyze time window is not set or the current time is not in the window, return false.
-func (a autoAnalysisTimeWindow) isWithinTimeWindow(currentTime time.Time) bool {
-	if a.start == (time.Time{}) || a.end == (time.Time{}) {
-		return false
-	}
-	return timeutil.WithinDayTimePeriod(a.start, a.end, currentTime)
 }
