+++ pkg/statistics/handle/autoanalyze/refresher/refresher_test.go
@@ -17,18 +17,15 @@ package refresher_test
 import (
 	"context"
 	"math"
-	"sort"
 	"testing"
 	"time"
 
-	"github.com/pingcap/tidb/pkg/meta/model"
 	pmodel "github.com/pingcap/tidb/pkg/parser/model"
 	"github.com/pingcap/tidb/pkg/statistics"
 	"github.com/pingcap/tidb/pkg/statistics/handle/autoanalyze/priorityqueue"
 	"github.com/pingcap/tidb/pkg/statistics/handle/autoanalyze/refresher"
 	"github.com/pingcap/tidb/pkg/testkit"
 	"github.com/stretchr/testify/require"
-	"github.com/tikv/client-go/v2/oracle"
 )
 
 func TestSkipAnalyzeTableWhenAutoAnalyzeRatioIsZero(t *testing.T) {
@@ -439,451 +436,4 @@ func TestRebuildTableAnalysisJobQueue(t *testing.T) {
 	require.Equal(t, float64(1), indicators.ChangePercentage)
 	require.Equal(t, float64(6*2), indicators.TableSize)
 	require.GreaterOrEqual(t, indicators.LastAnalysisDuration, time.Duration(0))
-}
-
-func TestCalculateChangePercentage(t *testing.T) {
-	unanalyzedColumns := map[int64]*statistics.Column{
-		1: {},
-		2: {},
-	}
-	unanalyzedIndices := map[int64]*statistics.Index{
-		1: {},
-		2: {},
-	}
-	analyzedColumns := map[int64]*statistics.Column{
-		1: {
-			StatsVer: 2,
-		},
-		2: {
-			StatsVer: 2,
-		},
-	}
-	analyzedIndices := map[int64]*statistics.Index{
-		1: {
-			StatsVer: 2,
-		},
-		2: {
-			StatsVer: 2,
-		},
-	}
-	bothUnanalyzedMap := statistics.NewColAndIndexExistenceMap(0, 0)
-	bothAnalyzedMap := statistics.NewColAndIndexExistenceMap(2, 2)
-	bothAnalyzedMap.InsertCol(1, nil, true)
-	bothAnalyzedMap.InsertCol(2, nil, true)
-	bothAnalyzedMap.InsertIndex(1, nil, true)
-	bothAnalyzedMap.InsertIndex(2, nil, true)
-	tests := []struct {
-		name             string
-		tblStats         *statistics.Table
-		autoAnalyzeRatio float64
-		want             float64
-	}{
-		{
-			name: "Test Table not analyzed",
-			tblStats: &statistics.Table{
-				HistColl:              *statistics.NewHistCollWithColsAndIdxs(0, false, statistics.AutoAnalyzeMinCnt+1, 0, unanalyzedColumns, unanalyzedIndices),
-				ColAndIdxExistenceMap: bothUnanalyzedMap,
-			},
-			autoAnalyzeRatio: 0.5,
-			want:             1,
-		},
-		{
-			name: "Based on change percentage",
-			tblStats: &statistics.Table{
-				HistColl:              *statistics.NewHistCollWithColsAndIdxs(0, false, statistics.AutoAnalyzeMinCnt+1, (statistics.AutoAnalyzeMinCnt+1)*2, analyzedColumns, analyzedIndices),
-				ColAndIdxExistenceMap: bothAnalyzedMap,
-				LastAnalyzeVersion:    1,
-			},
-			autoAnalyzeRatio: 0.5,
-			want:             2,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got := refresher.CalculateChangePercentage(tt.tblStats, tt.autoAnalyzeRatio)
-			require.Equal(t, tt.want, got)
-		})
-	}
-}
-
-func TestGetTableLastAnalyzeDuration(t *testing.T) {
-	// 2023-12-31 10:00:00
-	lastUpdateTime := time.Date(2023, 12, 31, 10, 0, 0, 0, time.UTC)
-	lastUpdateTs := oracle.GoTimeToTS(lastUpdateTime)
-	tblStats := &statistics.Table{
-		HistColl: *statistics.NewHistCollWithColsAndIdxs(0, false, 0, 0, map[int64]*statistics.Column{
-			1: {
-				StatsVer: 2,
-				Histogram: statistics.Histogram{
-					LastUpdateVersion: lastUpdateTs,
-				},
-			},
-		}, nil),
-		LastAnalyzeVersion: lastUpdateTs,
-	}
-	// 2024-01-01 10:00:00
-	currentTime := time.Date(2024, 1, 1, 10, 0, 0, 0, time.UTC)
-	currentTs := oracle.GoTimeToTS(currentTime)
-	want := 24 * time.Hour
-
-	got := refresher.GetTableLastAnalyzeDuration(tblStats, currentTs)
-	require.Equal(t, want, got)
-}
-
-func TestGetTableLastAnalyzeDurationForUnanalyzedTable(t *testing.T) {
-	tblStats := &statistics.Table{
-		HistColl: statistics.HistColl{},
-	}
-	// 2024-01-01 10:00:00
-	currentTime := time.Date(2024, 1, 1, 10, 0, 0, 0, time.UTC)
-	currentTs := oracle.GoTimeToTS(currentTime)
-	want := 1800 * time.Second
-
-	got := refresher.GetTableLastAnalyzeDuration(tblStats, currentTs)
-	require.Equal(t, want, got)
-}
-
-func TestCheckIndexesNeedAnalyze(t *testing.T) {
-	analyzedMap := statistics.NewColAndIndexExistenceMap(1, 0)
-	analyzedMap.InsertCol(1, nil, true)
-	analyzedMap.InsertIndex(1, nil, false)
-	tests := []struct {
-		name     string
-		tblInfo  *model.TableInfo
-		tblStats *statistics.Table
-		want     []string
-	}{
-		{
-			name: "Test Table not analyzed",
-			tblInfo: &model.TableInfo{
-				Indices: []*model.IndexInfo{
-					{
-						ID:    1,
-						Name:  pmodel.NewCIStr("index1"),
-						State: model.StatePublic,
-					},
-				},
-			},
-			tblStats: &statistics.Table{ColAndIdxExistenceMap: statistics.NewColAndIndexExistenceMap(0, 0)},
-			want:     nil,
-		},
-		{
-			name: "Test Index not analyzed",
-			tblInfo: &model.TableInfo{
-				Indices: []*model.IndexInfo{
-					{
-						ID:    1,
-						Name:  pmodel.NewCIStr("index1"),
-						State: model.StatePublic,
-					},
-				},
-			},
-			tblStats: &statistics.Table{
-				HistColl: *statistics.NewHistCollWithColsAndIdxs(0, false, 0, 0, map[int64]*statistics.Column{
-					1: {
-						StatsVer: 2,
-					},
-				}, map[int64]*statistics.Index{}),
-				ColAndIdxExistenceMap: analyzedMap,
-				LastAnalyzeVersion:    1,
-			},
-			want: []string{"index1"},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got := refresher.CheckIndexesNeedAnalyze(tt.tblInfo, tt.tblStats)
-			require.Equal(t, tt.want, got)
-		})
-	}
-}
-
-func TestCalculateIndicatorsForPartitions(t *testing.T) {
-	// 2024-01-01 10:00:00
-	currentTime := time.Date(2024, 1, 1, 10, 0, 0, 0, time.UTC)
-	currentTs := oracle.GoTimeToTS(currentTime)
-	// 2023-12-31 10:00:00
-	lastUpdateTime := time.Date(2023, 12, 31, 10, 0, 0, 0, time.UTC)
-	lastUpdateTs := oracle.GoTimeToTS(lastUpdateTime)
-	unanalyzedMap := statistics.NewColAndIndexExistenceMap(0, 0)
-	analyzedMap := statistics.NewColAndIndexExistenceMap(2, 1)
-	analyzedMap.InsertCol(1, nil, true)
-	analyzedMap.InsertCol(2, nil, true)
-	analyzedMap.InsertIndex(1, nil, true)
-	tests := []struct {
-		name                       string
-		globalStats                *statistics.Table
-		partitionStats             map[refresher.PartitionIDAndName]*statistics.Table
-		defs                       []model.PartitionDefinition
-		autoAnalyzeRatio           float64
-		currentTs                  uint64
-		wantAvgChangePercentage    float64
-		wantAvgSize                float64
-		wantAvgLastAnalyzeDuration time.Duration
-		wantPartitions             []string
-	}{
-		{
-			name: "Test Table not analyzed",
-			globalStats: &statistics.Table{
-				ColAndIdxExistenceMap: analyzedMap,
-			},
-			partitionStats: map[refresher.PartitionIDAndName]*statistics.Table{
-				{
-					ID:   1,
-					Name: "p0",
-				}: {
-					HistColl: statistics.HistColl{
-						Pseudo:        false,
-						RealtimeCount: statistics.AutoAnalyzeMinCnt + 1,
-					},
-					ColAndIdxExistenceMap: unanalyzedMap,
-				},
-				{
-					ID:   2,
-					Name: "p1",
-				}: {
-					HistColl: statistics.HistColl{
-						Pseudo:        false,
-						RealtimeCount: statistics.AutoAnalyzeMinCnt + 1,
-					},
-					ColAndIdxExistenceMap: unanalyzedMap,
-				},
-			},
-			defs: []model.PartitionDefinition{
-				{
-					ID:   1,
-					Name: pmodel.NewCIStr("p0"),
-				},
-				{
-					ID:   2,
-					Name: pmodel.NewCIStr("p1"),
-				},
-			},
-			autoAnalyzeRatio:           0.5,
-			currentTs:                  currentTs,
-			wantAvgChangePercentage:    1,
-			wantAvgSize:                2002,
-			wantAvgLastAnalyzeDuration: 1800 * time.Second,
-			wantPartitions:             []string{"p0", "p1"},
-		},
-		{
-			name: "Test Table analyzed and only one partition meets the threshold",
-			globalStats: &statistics.Table{
-				ColAndIdxExistenceMap: analyzedMap,
-			},
-			partitionStats: map[refresher.PartitionIDAndName]*statistics.Table{
-				{
-					ID:   1,
-					Name: "p0",
-				}: {
-					HistColl: *statistics.NewHistCollWithColsAndIdxs(0, false, statistics.AutoAnalyzeMinCnt+1, (statistics.AutoAnalyzeMinCnt+1)*2, map[int64]*statistics.Column{
-						1: {
-							StatsVer: 2,
-							Histogram: statistics.Histogram{
-								LastUpdateVersion: lastUpdateTs,
-							},
-						},
-						2: {
-							StatsVer: 2,
-							Histogram: statistics.Histogram{
-								LastUpdateVersion: lastUpdateTs,
-							},
-						},
-					}, nil),
-					Version:               currentTs,
-					ColAndIdxExistenceMap: analyzedMap,
-					LastAnalyzeVersion:    lastUpdateTs,
-				},
-				{
-					ID:   2,
-					Name: "p1",
-				}: {
-					HistColl: *statistics.NewHistCollWithColsAndIdxs(0, false, statistics.AutoAnalyzeMinCnt+1, 0, map[int64]*statistics.Column{
-						1: {
-							StatsVer: 2,
-							Histogram: statistics.Histogram{
-								LastUpdateVersion: lastUpdateTs,
-							},
-						},
-						2: {
-							StatsVer: 2,
-							Histogram: statistics.Histogram{
-								LastUpdateVersion: lastUpdateTs,
-							},
-						},
-					}, nil),
-					Version:               currentTs,
-					ColAndIdxExistenceMap: analyzedMap,
-					LastAnalyzeVersion:    lastUpdateTs,
-				},
-			},
-			defs: []model.PartitionDefinition{
-				{
-					ID:   1,
-					Name: pmodel.NewCIStr("p0"),
-				},
-				{
-					ID:   2,
-					Name: pmodel.NewCIStr("p1"),
-				},
-			},
-			autoAnalyzeRatio:           0.5,
-			currentTs:                  currentTs,
-			wantAvgChangePercentage:    2,
-			wantAvgSize:                2002,
-			wantAvgLastAnalyzeDuration: 24 * time.Hour,
-			wantPartitions:             []string{"p0"},
-		},
-		{
-			name: "No partition meets the threshold",
-			globalStats: &statistics.Table{
-				ColAndIdxExistenceMap: analyzedMap,
-			},
-			partitionStats: map[refresher.PartitionIDAndName]*statistics.Table{
-				{
-					ID:   1,
-					Name: "p0",
-				}: {
-					HistColl: *statistics.NewHistCollWithColsAndIdxs(0, false, statistics.AutoAnalyzeMinCnt+1, 0, map[int64]*statistics.Column{
-						1: {
-							StatsVer: 2,
-							Histogram: statistics.Histogram{
-								LastUpdateVersion: lastUpdateTs,
-							},
-						},
-						2: {
-							StatsVer: 2,
-							Histogram: statistics.Histogram{
-								LastUpdateVersion: lastUpdateTs,
-							},
-						},
-					}, nil),
-					Version:               currentTs,
-					ColAndIdxExistenceMap: analyzedMap,
-					LastAnalyzeVersion:    lastUpdateTs,
-				},
-				{
-					ID:   2,
-					Name: "p1",
-				}: {
-					HistColl: *statistics.NewHistCollWithColsAndIdxs(0, false, statistics.AutoAnalyzeMinCnt+1, 0, map[int64]*statistics.Column{
-						1: {
-							StatsVer: 2,
-							Histogram: statistics.Histogram{
-								LastUpdateVersion: lastUpdateTs,
-							},
-						},
-						2: {
-							StatsVer: 2,
-							Histogram: statistics.Histogram{
-								LastUpdateVersion: lastUpdateTs,
-							},
-						},
-					}, nil),
-					Version:               currentTs,
-					ColAndIdxExistenceMap: analyzedMap,
-					LastAnalyzeVersion:    lastUpdateTs,
-				},
-			},
-			defs: []model.PartitionDefinition{
-				{
-					ID:   1,
-					Name: pmodel.NewCIStr("p0"),
-				},
-				{
-					ID:   2,
-					Name: pmodel.NewCIStr("p1"),
-				},
-			},
-			autoAnalyzeRatio:           0.5,
-			currentTs:                  currentTs,
-			wantAvgChangePercentage:    0,
-			wantAvgSize:                0,
-			wantAvgLastAnalyzeDuration: 0,
-			wantPartitions:             []string{},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			gotAvgChangePercentage,
-				gotAvgSize,
-				gotAvgLastAnalyzeDuration,
-				gotPartitions :=
-				refresher.CalculateIndicatorsForPartitions(
-					tt.globalStats,
-					tt.partitionStats,
-					tt.autoAnalyzeRatio,
-					tt.currentTs,
-				)
-			require.Equal(t, tt.wantAvgChangePercentage, gotAvgChangePercentage)
-			require.Equal(t, tt.wantAvgSize, gotAvgSize)
-			require.Equal(t, tt.wantAvgLastAnalyzeDuration, gotAvgLastAnalyzeDuration)
-			// Sort the partitions.
-			sort.Strings(tt.wantPartitions)
-			sort.Strings(gotPartitions)
-			require.Equal(t, tt.wantPartitions, gotPartitions)
-		})
-	}
-}
-
-func TestCheckNewlyAddedIndexesNeedAnalyzeForPartitionedTable(t *testing.T) {
-	tblInfo := model.TableInfo{
-		Indices: []*model.IndexInfo{
-			{
-				ID:    1,
-				Name:  pmodel.NewCIStr("index1"),
-				State: model.StatePublic,
-			},
-			{
-				ID:    2,
-				Name:  pmodel.NewCIStr("index2"),
-				State: model.StatePublic,
-			},
-		},
-		Columns: []*model.ColumnInfo{
-			{
-				ID: 1,
-			},
-			{
-				ID: 2,
-			},
-		},
-	}
-	partitionStats := map[refresher.PartitionIDAndName]*statistics.Table{
-		{
-			ID:   1,
-			Name: "p0",
-		}: {
-			HistColl:              *statistics.NewHistCollWithColsAndIdxs(0, false, statistics.AutoAnalyzeMinCnt+1, 0, nil, map[int64]*statistics.Index{}),
-			ColAndIdxExistenceMap: statistics.NewColAndIndexExistenceMap(0, 0),
-		},
-		{
-			ID:   2,
-			Name: "p1",
-		}: {
-			HistColl: *statistics.NewHistCollWithColsAndIdxs(0, false, statistics.AutoAnalyzeMinCnt+1, 0, nil, map[int64]*statistics.Index{
-				2: {
-					StatsVer: 2,
-				},
-			}),
-			ColAndIdxExistenceMap: statistics.NewColAndIndexExistenceMap(0, 1),
-		},
-	}
-
-	partitionIndexes := refresher.CheckNewlyAddedIndexesNeedAnalyzeForPartitionedTable(&tblInfo, partitionStats)
-	expected := map[string][]string{"index1": {"p0", "p1"}, "index2": {"p0"}}
-	require.Equal(t, len(expected), len(partitionIndexes))
-
-	for k, v := range expected {
-		sort.Strings(v)
-		if val, ok := partitionIndexes[k]; ok {
-			sort.Strings(val)
-			require.Equal(t, v, val)
-		} else {
-			require.Fail(t, "key not found in partitionIndexes: "+k)
-		}
-	}
 }
